requirements.txt
databricks-sdk>=0.18.0
tenacity>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0


